tests for sms

+++ sms.go
@@ -6,7 +6,7 @@ import (
 )
 
 // SmsCalculateSmsFragments calculate based on a string how many SMS messages will it take to deliver a massage
-func SmsCalculateSmsFragments(message string) uint64 {
+func SmsCalculateSmsFragments(message string) uint16 {
 	mbLen := utf8.RuneCountInString(message)
 	byeLen := len(message)
 	maxMessageLength := 160
@@ -18,7 +18,7 @@ func SmsCalculateSmsFragments(message string) uint64 {
 	}
 
 	if mbLen > maxMessageLength {
-		return uint64(math.Ceil(float64(mbLen) / float64(maxMultiMessageLength)))
+		return uint16(math.Ceil(float64(mbLen) / float64(maxMultiMessageLength)))
 	}
 
 	return 1
